🏷️ Split types into separate files
