+++ Makefile
@@ -29,6 +29,5 @@ prod:	## Build for Production environment
 publish:	## Publish to NPM
 	@make prod
 	@pnpm copy-dts-declarations
-	@pnpm duplicate-dts
 	@make tests
 	@pnpm publish
